[FEATURE] Add password hashing
In the extension configuration can be defined, how a password should be
hashed. Plaintext passwords are not supported. Available hashing
methods:

1. Bcrypt (default)
   Standard bcrypt hashing, like it is supported by lots of systems. To
   balance speed/security it is possible to configure the 'cost'
   parameter of bcrypt.
2. TYPO3 password hashing
   Uses the password hashing facilities from TYPO3. The hashing
   instances are created for FE mode (configuration from
   $GLOBALS['TYPO3_CONF_VARS']['FE'']['passwordHashing'] applies).
   Although this method can be much more secure, it will come at the
   cost of performance, because **every request** must be authenticated
   (there is no state or session)!

The support for the default password hash of the htpasswd program
(Apache-style MD5 salted hashing $apr1$) will be added in the future.

Regardless of the configuration to generate password hashes, the
verification process will work with any supported hashing method (by
this extension or TYPO3).